Fix pool shuffling (#145)
* Fix pool shuffling

When connecting to a server, the received INFO may contain an array
of URLs. The current code was adding those and then shuffeling the
entire pool (if NoRandomize is false). This is wrong since the
connect (and reconnect) loop would then potentially skip never
tried servers and try again servers that had just failed.

* Only randomize discovered servers
* Refactored the add apis in server pool (return value is no longer needed)

*

* Additional checks allowing the cluster to form.

* Fix flapping tests (running on slow machines).

* Add more time for cluster formation.

* Further shore up tests, allow longer time for cluster formation.

* Utilize the same url randomization for initial connect and randomizing discovered servers.