Implements ServerDiscoveredEventHandler from #159
ServerPool.add matches its description and returns a boolean indicating if the server url
is new for the pool. Connection.processInfoArgs uses this to signal on the new event if and
only if it has been asked to notify on server discovery. The initial connect path does not enable
that behavior, but the processAsyncInfo path does enable it, matching the behavior of the
golang client.