Prevent the ConnectAsync task from generating unobserved task exceptions.
If the ConnectAsync task threw an exception after the Wait-call had returned, it was never observed, triggering TaskScheduler.UnobservedTaskException when the task was garbage collected. Fixed by adding a dummy continuation which does nothing but observing the potential exception. The only call site for this method is createConn, which silences all exceptions, so it seems safe to do so here too.