Remove lock in EncodedConn.publishObject
There is no reason to lock here when the only meaningful part is to call
Connection.publish(), which itself can be called with the exact same
arguments through Connection.Publish().

Moreover, holding the lock throughout the publication prevents a successful
reconnect (which in practice will try to acquire it), so that the publication
is guaranteed to fail even if it could have succeeded after reconnection.