Fix swallowed exceptions
Following dde17f20, the waited Task would never cause task.Wait() to throw :
In case of connection failure, the code will continue and throw later at
client.GetStream(), with a much less meaningful exception.

We should therefore wait for the real connection task, and let the
continuation only be useful in case of a timeout.