Fix line reader to resolve buffer issues on TLS connections
The StreamReader usage when reading the server INFO causes buffer
issues on the stream, as it is also read in the reader loop. The
StreamReader might consume or return the read bytes back, leading to
problems when used on a TLS connection with .NET versions lower than 8.
This fix removes the StreamReader usage to address these issues.